NPC updates
Fixed up some dialogue functions to deal with the new format.

Finished implementing relationships; at least a first pass. Affinity updates as you give them items and is persistent. Dialogue choices based on affinity change to match the current affinity as it updates.

+++ Hocus Potions/Assets/Scripts/NPCManager.cs
@@ -20,7 +20,7 @@ public class NPCManager : MonoBehaviour {
         public int returningDay;
         public int returningHour;
         public int returningMinutes;
-        public int affinity;
+        public float affinity;
     }
 
     public Dictionary<string, NPCData> data;
